+++ inscription_aux_cours/views/cours/formulaire.py
@@ -38,6 +38,7 @@ from osis_learning_unit_sdk.model.classe import Classe
 from osis_learning_unit_sdk.model.learning_unit import LearningUnit
 
 from base.services.utils import ServiceException
+from base.utils.string_utils import unaccent
 from inscription_aux_cours.forms.cours.demande_particuliere import DemandeParticuliereForm
 from inscription_aux_cours.forms.cours.inscription_hors_programme import InscriptionHorsProgrammeForm
 from inscription_aux_cours.services.cours import CoursService
@@ -74,6 +75,7 @@ class FormulaireInscriptionAuxCoursView(LoginRequiredMixin,  InscriptionAuxCours
             for formulaire_mini_formation in self.formulaire_inscriptions_cours.formulaires_mini_formation
             if formulaire_mini_formation.configuration.autorise_etudiant_a_ajouter_cours
         ]
+        choix_mini_formations.sort(key=lambda choix: unaccent(choix[1]))
         if self.formulaire_inscriptions_cours.formulaire_tronc_commun.configuration.autorise_etudiant_a_ajouter_cours:
             choix_mini_formations = [('', self.programme.intitule_formation)] + choix_mini_formations
         return InscriptionHorsProgrammeForm(
